requirements.txt
opencv-python>=4.8
numpy>=1.24
torch>=2.0
torchvision>=0.15
mmengine>=0.9
mmcv-lite>=2.0,<2.2
mmdet>=3.0
mmaction2>=1.0
emotiefflib>=1.1
ultralytics>=8.0
requests>=2.28  # For robust EmotiEffLib model download (avoids 429)
PyYAML>=6.0
fastapi>=0.110
uvicorn>=0.29
pytest>=8.0
# Person detection uses YOLOv8 (ultralytics) by default.
# mmaction2/emotiefflib may have optional deps; action/emotion degrade gracefully if unavailable.
